requirements.txt
google-cloud-vision
tweepy
pytest
Flask
requests